// types/iEvent.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface iEvent {
  eventName: string;
  /**
   * Date in this case refers to the date and time
   */
  date: {
    [k: string]: any;
  };
  place: string;
  room?: {
    [k: string]: any;
  };
}